+++ CHANGELOG.md
@@ -3,6 +3,7 @@ UNRELEASED
 
 * [ [#1664](https://github.com/digitalfabrik/integreat-cms/issues/1664) ] Fix media library and content edit lock in sbs view
 * [ [#1660](https://github.com/digitalfabrik/integreat-cms/issues/1660) ] Fix moving pages to the root level from the page form
+* [ [#1566](https://github.com/digitalfabrik/integreat-cms/issues/1566) ] Manage organizations per region
 
 
 2022.8.3